refactor: update clang-tidy checks and improve code readability
- Removed several clang-tidy checks that were deemed unnecessary or caused false positives, including narrowing conversions and function cognitive complexity.
- Adjusted formatting and alignment in various source files for improved readability.
- Updated function signatures to use const references where applicable, enhancing performance and clarity.
- Refined lambda expressions in ROS2 demo and bag recorder for better consistency.

+++ src/core/layout.cpp
@@ -35,7 +35,7 @@ std::vector<Rect> compute_subplot_layout(float          figure_width,
                                          float*         actual_content_height)
 {
     std::vector<Rect> rects;
-    rects.reserve(static_cast<size_t>(rows * cols));
+    rects.reserve(static_cast<size_t>(rows) * static_cast<size_t>(cols));
 
     // Total available space for the grid of subplots
     // Each cell gets an equal share of the figure, then margins are applied inside each cell.
